requirements.txt
python-telegram-bot~=8.1
SQLAlchemy~=1.1.4
sqlalchemy-utils~=0.32.12

# Logging
raven

# Linting
flake8
flake8-docstrings
flake8-print
flake8-commas
flake8-bugbear
flake8-string-format
